+++ crates/services/gas_price_service/src/common/gas_price_algorithm.rs
@@ -1,13 +1,14 @@
+use fuel_core_services::seqlock::SeqLock;
 use fuel_core_types::fuel_types::BlockHeight;
 use std::sync::Arc;
 
-pub trait GasPriceAlgorithm {
+pub trait GasPriceAlgorithm: Copy {
     fn next_gas_price(&self) -> u64;
     fn worst_case_gas_price(&self, block_height: BlockHeight) -> u64;
 }
 
 #[derive(Debug, Default)]
-pub struct SharedGasPriceAlgo<A>(Arc<parking_lot::RwLock<A>>);
+pub struct SharedGasPriceAlgo<A>(Arc<SeqLock<A>>);
 
 impl<A> Clone for SharedGasPriceAlgo<A> {
     fn clone(&self) -> Self {
@@ -20,12 +21,13 @@ where
     A: Send + Sync,
 {
     pub fn new_with_algorithm(algorithm: A) -> Self {
-        Self(Arc::new(parking_lot::RwLock::new(algorithm)))
+        Self(Arc::new(SeqLock::new(algorithm)))
     }
 
-    pub async fn update(&mut self, new_algo: A) {
-        let mut write_lock = self.0.write();
-        *write_lock = new_algo;
+    pub fn update(&mut self, new_algo: A) {
+        self.0.write(|data| {
+            *data = new_algo;
+        });
     }
 }
 
@@ -37,7 +39,7 @@ where
         self.0.read().next_gas_price()
     }
 
-    pub async fn worst_case_gas_price(&self, block_height: BlockHeight) -> u64 {
+    pub fn worst_case_gas_price(&self, block_height: BlockHeight) -> u64 {
         self.0.read().worst_case_gas_price(block_height)
     }
 }